requirements.txt
discord.py[voice]>=2.4.0
requests>=2.32.3
aiohttp>=3.9.5
yt-dlp>=2025.1.26
ffmpeg-python>=0.2.0
PyNaCl>=1.5.0
pytz